Another revamp of NVPair & test improvements.
Add a basic functional test which essentially detects an available drive to
create a pool on (FreeBSD-specific), creates a pool, and generates a
ZFS::Pool object.

Revamp NVPair so that values are now represented as NVValues.  This enables
us to create NVValue::Array objects which are represented as ruby-like Array
objects containing NVValues.