Fix NVValue::String#to_native, which is a **.
This fixes the unit test for the string nvpair.  This was previously broken
because I noticed that strings were not parsed correctly in reading the
nvlist for a vdev.  That was fixed in 10200bc238cc819f4fc2c2709ea78fcdab177b2d,
so this is just to fix the unit tests and NVValue::String#to_native itself.