Fix a chicken and egg require problem.
It raises an exception anytime somebody requires this gem, including its
own tests.  The gem works on some systems and not others because
Dir.each does not have a deterministic order, and we were sensitive to
the order of the requires.

lib/zfs/nvvalue.rb Always require "array" before any of the other files.